Removed overflow hidden custom hook
